+++ README.md
@@ -1,1 +1,2 @@
 # node-auth-jwt
+Node, express, mongoDB, mongoose, passport and JWT REST API authentication
