+++ src/config/main.js
